requirements.txt
backports.csv==1.0.7
beautifulsoup4==4.7.1
construct==2.5.3
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.2
django-htmlmin==0.11.0
django-import-export==1.2.0
et-xmlfile==1.0.1
future==0.17.1
html5lib==1.0.1
jdcal==1.4
mysqlclient==1.4.2
odfpy==1.4.0
openpyxl==2.6.1
pefile==2018.8.8
python-ptrace==0.9.3
pytz==2018.9
PyYAML==5.1
six==1.12.0
soupsieve==1.8
sqlparse==0.3.0
tablib==0.13.0
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
